Basic updates
removed the lorem ipsum articles for now as they weren't adding anything. added a (currently) non functioning nav bar to the project. Added a red variant to the quiz answers so that if a user gets a question wrong its highlighted in red. added the us and france hints to the first question.