LOTS OF CHANGES
big apologies, i did a lot of changes and forgot to push. (i am very sorry) essentially the quiz is now all on the page to begin with so we can intertwine 'articles' with our questions. we've added some ACTUAL companies gender pay gap data instead of job roles to make it more personal and impactful. changed the form of the match the company and the pay gap as well so that its more of a 'pairs' style game with cute colours to add some more colour to our work. need to work on implementing the about us page now!!! (also realised my 'about' page was spelt wrong that's fixed now)