Final Update
The final major updates to the code. this was fixing the answering mechanics (essentially if you answered a question incorrect and then the next one correct it made it say 'correct' on the one that was incorrect. added a reference for the second question. removed the women.png image (didn't have the licensing for it)